refactor(database): several sql cleanups and query fixes
